backend/requirements.txt
Django>=4.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
Pillow
psycopg2-binary
django-knox
